add handler for add

+++ internal/account/database/model/goods.go
@@ -5,5 +5,5 @@ import "gorm.io/gorm"
 type GoodsPrice struct {
 	gorm.Model
 	Name  string `gorm:"type:varchar(128);unique;not null"`
-	Price int    `gorm:"not null"`
+	Price int64  `gorm:"not null"`
 }
